backend/requirements.txt
fastapi
uvicorn[standard]
pyodbc
python-multipart
python-dotenv
openpyxl
bcrypt
PyJWT
